requirements.txt
anyio
bcrypt
Body
bson
cffi
click
colorama
dnspython
ecdsa
email-validator
fastapi
fastapi-jwt-auth
h11
httpcore
httptools
httpx
idna
jwt
motor
rsa
passlib
pyasn1
pycparser
pydantic
PyJWT
pymongo
python-dotenv
python-jose
python-multipart
PyYAML
six
sniffio
starlette
uvicorn
